Update supl-ddf-file.md
We generalize SUPL version by adding FullVersion field, in consideration of backward compatibility. The desciption PR is out separately and we update SUPL DDF with this PR. This change applies from Vibranium release and we just add a new field, so it is backward compatibile. 

Tested with DDF Validator tool.